configurations entity added

+++ DataAccessLayer/Entities/Order.cs
@@ -7,7 +7,7 @@ namespace DataAccessLayer.Entities
 
         public DateTime OrderDate { get; set; }
         public float TotalAmount { get; set; }
-        public int DiscountAmount { get; set; }
+        public float DiscountAmount { get; set; }
 
         public long? UserId { get; set; }
         public User User { get; set; }
